+++ r.slopeunits.create/r.slopeunits.create.py
@@ -2,7 +2,7 @@
 #
 ############################################################################
 #
-# MODULE:       r.slopeunits for GRASS 7
+# MODULE:       r.slopeunits.create for GRASS 7
 # AUTHOR(S):    Ivan Marchesini, Massimiliano Alvioli
 # PURPOSE:      To create a raster layer of slope units
 # COPYRIGHT:    (C) 2004-2012 by the GRASS Development Team
@@ -36,12 +36,6 @@
 # % key: slumap
 # % description: Output Slope Units layer (the main output)
 # % required : yes
-# %end
-
-# %option G_OPT_R_OUTPUT
-# % key: slumapclean
-# % description: Output Slope Units layer, cleaned (the main output)
-# % required : no
 # %end
 
 # %option G_OPT_R_OUTPUT
@@ -96,25 +90,6 @@
 # % type: integer
 # % description: maximum number of iteration to do before the procedure is in any case stopped
 # % required : yes
-# %end
-
-# %option
-# % key: cleansize
-# % type: double
-# % description: Slope Units size to be removed
-# % required : no
-# %end
-
-# %flag
-# % key: m
-# % description: Perform quick cleaning of small-sized areas and stripes
-# % guisection: flags
-# %end
-
-# %flag
-# % key: n
-# % description: Perform detailed cleaning of small-sized areas (slow)
-# % guisection: flags
 # %end
 
 import atexit
@@ -725,728 +700,6 @@ def slope_units(
     grass.message("Slope units calculated.")
 
 
-def clean_method_3(input_vect, output_vect, minarea):
-    """Clean up"""
-    region = grass.region()
-    nsres = region["nsres"]
-    ewres = region["ewres"]
-    smarea = 10 * nsres * ewres
-
-    grass.run_command(
-        "v.clean",
-        input=input_vect,
-        output="slu_clean",
-        tool="rmarea",
-        threshold=smarea,
-        quiet=True,
-    )
-    grass.run_command(
-        "v.db.addcolumn",
-        map="slu_clean",
-        columns="area integer, perimetro integer",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.to.db", map="slu_clean", option="area", columns="area", quiet=True
-    )
-    grass.run_command(
-        "v.to.db",
-        map="slu_clean",
-        option="perimeter",
-        columns="perimetro",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.db.droprow",
-        input="slu_clean",
-        where="area is null",
-        output="slu_area",
-        quiet=True,
-    )
-    rm_vectors.append("slu_clean")
-
-    key = grass.vector_db(map="slu_area")[1]["key"]
-    lista = grass.read_command(
-        "v.db.select",
-        map="slu_area",
-        columns=key,
-        where=f"area <= {minarea}",
-        flags="c",
-    )
-    lista = lista.splitlines()
-    buchi = ",".join(lista)
-    totalebuchi = len(lista)
-    grass.run_command(
-        "v.extract",
-        input="slu_area",
-        cats=buchi,
-        output="slu_buchi",
-        type="area",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.to.rast",
-        input="slu_buchi",
-        output="slu_buchi",
-        use="cat",
-        overwrite=True,
-    )
-
-    key = grass.vector_db(map="slu_area")[1]["key"]
-    lista = grass.read_command(
-        "v.db.select",
-        map="slu_area",
-        columns=key,
-        where=f"area > {minarea}",
-        flags="c",
-    )
-    lista = lista.splitlines()
-    nobuchi = ",".join(lista)
-    grass.run_command(
-        "v.extract",
-        input="slu_area",
-        cats=nobuchi,
-        output="slu_nobuchi",
-        type="area",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.to.rast",
-        input="slu_nobuchi",
-        output="slu_nobuchi",
-        use="cat",
-        overwrite=True,
-    )
-    grass.run_command(
-        "v.category",
-        input="slu_area",
-        output="slu_bordi",
-        layer=2,
-        type="boundary",
-        option="add",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.db.addtable",
-        map="slu_bordi",
-        layer=2,
-        columns="left integer,right integer,lunghezza integer",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.to.db",
-        map="slu_bordi",
-        option="sides",
-        columns="left,right",
-        layer=2,
-        type="boundary",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.to.db",
-        map="slu_bordi",
-        option="length",
-        columns="lunghezza",
-        layer=2,
-        type="boundary",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.to.rast", input="slu_area", output="slu_area", use="cat", quiet=True
-    )
-    # TODO: different names than coseno and seno, these rasters are already
-    # created
-    grass.mapcalc("coseno = cos(aspect_slu)", quiet=True)
-    grass.mapcalc("seno = sin(aspect_slu)", quiet=True)
-
-    grass.run_command(
-        "r.stats.zonal",
-        base="slu_area",
-        cover="coseno",
-        method="count",
-        output="count",
-        quiet=True,
-    )
-    grass.run_command(
-        "r.stats.zonal",
-        base="slu_area",
-        cover="coseno",
-        method="sum",
-        output="cumcos",
-        quiet=True,
-    )
-    grass.run_command(
-        "r.stats.zonal",
-        base="slu_area",
-        cover="seno",
-        method="sum",
-        output="sumsin",
-        quiet=True,
-    )
-
-    grass.mapcalc("cos_medio = sumcos / count", quiet=True)
-    grass.mapcalc("sin_medio = sumsin / count", quiet=True)
-    grass.run_command(
-        "r.to.vect",
-        input="cos_medio",
-        output="cos_medio",
-        type="area",
-        quiet=True,
-    )
-    grass.run_command(
-        "r.to.vect",
-        input="sin_medio",
-        output="sin_medio",
-        type="area",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.overlay",
-        ainput="slu_area",
-        binput="cos_medio",
-        operator="and",
-        atype="area",
-        btype="area",
-        output="cos_medio_over",
-        quiet=True,
-    )
-    grass.run_command(
-        "v.overlay",
-        ainput="slu_area",
-        binput="sin_medio",
-        operator="and",
-        atype="area",
-        btype="area",
-        output="sin_medio_over",
-        quiet=True,
-    )
-
-    pulire = grass.read_command(
-        "v.category", input="slu_buchi", option="print", quiet=True
-    )
-    pulire = pulire.splitlines()
-
-    grass.run_command("g.copy", vector=f"slu_area,{output_vect}", quiet=True)
-
-    ico = 1
-    for i in pulire:
-        inti = int(i)
-        lista1 = grass.read_command(
-            "db.select",
-            sql=(
-                "select b2.right from slu_bordi_2 b2 "
-                f"where b2.left = {i} and b2.right <> -1"
-            ),
-            flags="c",
-        )
-        lista2 = grass.read_command(
-            "db.select",
-            sql=(
-                "select b2.left from slu_bordi_2 b2 "
-                f"where b2.left <> -1 and b2.right = {inti}"
-            ),
-            flags="c",
-        )
-        vicini = lista1.splitlines()
-        vicini.extend(lista2.splitlines)
-        vicini = sorted(set(vicini))
-        if len(vicini) > 0:
-            grass.message(
-                f" --- --- -- buco numero {ico} di {totalebuchi}, "
-                "cat: {i}, vicini: {vicini}"
-            )
-            ico = ico + 1
-            grass.message(f"vicini: {vicini}")
-            grass.run_command(
-                "v.extract",
-                input=output_vect,
-                cats=",".join(vicini),
-                output="intorno",
-                type="area",
-                quiet=True,
-            )
-            chk_intorno = grass.read_command(
-                "v.category",
-                input="intorno",
-                type="centroid",
-                option="print",
-                quiet=True,
-            )
-            chk_intorno = chk_intorno.splitlines()
-            if len(chk_intorno) > 0:
-                # potrei voler cambiare questo perche' quando ci sono buchi
-                # contigui fa un po' di casino
-                grass.run_command(
-                    "v.overlay",
-                    ainput="intorno",
-                    binput="slu_nobuchi",
-                    output="intorno_OK",
-                    atype="area",
-                    btype="area",
-                    olayer="0,1,0",
-                    operator="and",
-                    quiet=True,
-                )
-
-                cos_buco = grass.read_command(
-                    "v.db.select",
-                    map="cos_medio_over",
-                    where=f"a_cat={i}",
-                    columns="b_value",
-                    flags="c",
-                    quiet=True,
-                )
-                sin_buco = grass.read_command(
-                    "v.db.select",
-                    map="sin_medio_over",
-                    where=f"a_cat={i}",
-                    columns="b_value",
-                    flags="c",
-                    quiet=True,
-                )
-                grass.message(f"buco cat {i}: cos={cos_buco} sin={sin_buco}")
-                massimo = -10000
-                jmax = 0
-                loop = grass.read_command(
-                    "v.category",
-                    input="intorno_OK",
-                    option="print",
-                    quiet=True,
-                )
-                loop = loop.splitlines()
-                for j in loop:
-                    j = int(j)
-                    cos_j = grass.read_command(
-                        "v.db.select",
-                        map="cos_medio_over",
-                        where=f"a_cat={j}",
-                        columns="b_value",
-                        flags="c",
-                        quiet=True,
-                    )
-                    sin_j = grass.read_command(
-                        "v.db.select",
-                        map="sin_medio_over",
-                        where=f"a_cat={j}",
-                        columns="b_value",
-                        flags="c",
-                        quiet=True,
-                    )
-                    dotpr = (
-                        float(cos_buco) * float(cos_j)
-                        + float(sin_buco) * float(sin_j)
-                    ) * 10000
-                    if dotpr >= massimo and dotpr > 0:
-                        massimo = dotpr
-                        jmax = j
-
-                    grass.message(
-                        f"i: {i} j: {j} cos_j: {cos_j} sin_j: {sin_j} "
-                        "dotpr: {dotpr} jmax: {jmax}"
-                    )
-
-                grass.message(f"massimo: {massimo} per j={jmax}")
-                if jmax > 0:
-                    lunghezza = grass.read_command(
-                        "db.select",
-                        sql=(
-                            "select b2.lunghezza from slu_bordi_2 b2 where "
-                            f"(b2.left={i} and b2.right={jmax}) "
-                            f"or (b2.left={jmax} and b2.right={i})"
-                        ),
-                        flags="c",
-                        quiet=True,
-                    )
-                    perimetro = grass.read_command(
-                        "v.db.select",
-                        map="slu_clean",
-                        columns="perimetro",
-                        where=f"cat={i}",
-                        flags="c",
-                        quiet=True,
-                    )
-                    lunghezza = float(lunghezza)
-                    perimetro = float(perimetro)
-                    if lunghezza > 0 and perimetro > 0:
-                        frazione = lunghezza / perimetro * 10000
-                        if frazione > 500:
-                            grass.message(
-                                f"lungh: {lunghezza}; perim: {perimetro}; "
-                                f"fract: {frazione}"
-                            )
-                            grass.run_command(
-                                "v.extract",
-                                input=output_vect,
-                                output="slu_i",
-                                cats=f"{i},{jmax}",
-                                new={jmax},
-                                flags="d",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.overlay",
-                                ainput=output_vect,
-                                binput="slu_i",
-                                atype="area",
-                                btype="area",
-                                operator="not",
-                                olayer="0,1,0",
-                                output="slu_j",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.overlay",
-                                ainput="slu_i",
-                                binput="slu_j",
-                                atype="area",
-                                btype="area",
-                                operator="or",
-                                output="slu_k",
-                                olayer="1,0,0",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.db.addcolumn",
-                                map="slu_k",
-                                column="newcat integer",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.db.update",
-                                map="slu_k",
-                                layer=1,
-                                column="newcat",
-                                qcolumn="a_cat",
-                                where="a_cat is not null",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.db.update",
-                                map="slu_k",
-                                layer=1,
-                                column="newcat",
-                                qcolumn="b_cat",
-                                where="b_cat is not null",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.reclass",
-                                input="slu_k",
-                                output=output_vect,
-                                column="newcat",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.db.addtable", map=output_vect, quiet=True
-                            )
-                            grass.run_command(
-                                "v.db.addcolumn",
-                                map=output_vect,
-                                columns="area integer",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "v.to.db",
-                                map=output_vect,
-                                option="area",
-                                columns="area",
-                                quiet=True,
-                            )
-                            grass.run_command(
-                                "g.remove",
-                                type="vector",
-                                name="slu_i,slu_j,slu_k",
-                                flags="f",
-                                quiet=True,
-                            )
-                    # lunghezza and perimetro
-                # jmax
-                grass.run_command(
-                    "g.remove",
-                    type="vector",
-                    name="intorno_OK",
-                    flags="f",
-                    quiet=True,
-                )
-            # chk_category
-            grass.run_command(
-                "g.remove",
-                type="vector",
-                name="intorno",
-                flags="f",
-                quiet=True,
-            )
-        # vicini
-    # pulire
-
-
-def clean_small_areas(dem, slumap, plains, cleansize=-1, slumapclean=None):
-    """Cleaning of small areas"""
-    region = grass.region()
-    nsres = region["nsres"]
-    ewres = region["ewres"]
-
-    if not flags["n"]:
-        if not flags["m"]:
-            grass.message(
-                " -- we want QUICK cleaning of small-sized areas: METHOD 1 --"
-            )
-
-        exp = "$out = if(isnull($mask), null(), 1)"
-        grass.mapcalc(exp, out="MASK", mask=dem, quiet=True)
-        areamap = "areamap"
-
-        grass.run_command(
-            "r.clump", input=slumap, output="slu_clump", quiet=True
-        )
-        rm_rasters.append("slu_clump")
-
-        grass.run_command(
-            "r.stats.zonal",
-            base="slu_clump",
-            cover="slu_clump",
-            method="count",
-            output="slu_count",
-            quiet=True,
-        )
-        rm_rasters.append("slu_count")
-
-        exp = "$out = $a * $b * $c"
-        grass.mapcalc(
-            exp,
-            out=areamap,
-            a="slu_count",
-            b=nsres,
-            c=ewres,
-            quiet=True,
-        )
-        rm_rasters.append(areamap)
-
-        exp = "$out = if($a > $b, $c, null())"
-        grass.mapcalc(
-            exp,
-            out="slu_r_clean",
-            a=areamap,
-            b=cleansize,
-            c=slumap,
-            quiet=True,
-        )
-        rm_rasters.append("slu_r_clean")
-
-        cleansize = cleansize / (nsres * ewres)
-        growdist = int((10 * cleansize / 3.14) ** 0.5)
-        grass.run_command(
-            "r.grow",
-            input="slu_r_clean",
-            output="slu_r_grow",
-            radius=growdist,
-            quiet=True,
-        )
-        rm_rasters.append("slu_r_grow")
-
-    if flags["m"]:
-        grass.message(
-            " -- we want QUICK cleaning of small-sized areas: METHOD 2 --"
-        )
-        clean_input = "slu_r_grow"
-        clean_output = "slu_no_stripes"
-        grass.run_command(
-            "r.neighbors",
-            input=clean_input,
-            output="slu_diversity",
-            method="diversity",
-            size=5,
-            quiet=True,
-        )
-        rm_rasters.append("slu_diversity")
-
-        exp = "$out = if($a == 1, 1, null())"
-        grass.mapcalc(
-            exp,
-            out="slu_diversity_nobordi",
-            a="slu_diversity",
-            quiet=True,
-        )
-        rm_rasters.append("slu_diversity_nobordi")
-
-        grass.run_command(
-            "r.grow",
-            input="slu_diversity_nobordi",
-            output="slu_diversity_nobordi_grow",
-            radius=1.01,
-            quiet=True,
-        )
-        rm_rasters.append("slu_diversity_nobordi_grow")
-
-        exp = "$out = if(isnull($a), null(), $b)"
-        grass.mapcalc(
-            exp,
-            out="slu_finale_nobordi",
-            a="slu_diversity_nobordi_grow",
-            b=clean_input,
-            quiet=True,
-        )
-        rm_rasters.append("slu_finale_nobordi")
-
-        # TODO: fixed radius ?
-        grass.run_command(
-            "r.grow",
-            input="slu_finale_nobordi",
-            output=clean_output,
-            radius=1000,
-            quiet=True,
-        )
-        rm_rasters.append(clean_output)
-
-        exp = "$out = int($a)"
-        # output="slu_r_grow" exists already
-        grass.run_command(
-            "g.remove", type="raster", name=clean_input, flags="f", quiet=True
-        )
-        grass.mapcalc(exp, out=clean_input, a=clean_output, quiet=True)
-        grass.run_command(
-            "g.remove",
-            type="raster",
-            name=(
-                "slu_diversity,slu_diversity_nobordi,"
-                "slu_diversity_nobordi_grow,slu_finale_nobordi,slu_no_stripes"
-            ),
-            flags="f",
-            quiet=True,
-        )
-
-    if flags["n"]:
-        grass.message(
-            " -- we want DETAILED cleaning of small-sized areas: METHOD 3 --"
-        )
-        grass.run_command(
-            "r.to.vect",
-            input=slumap,
-            output="slu_v_grow",
-            type="area",
-            quiet=True,
-        )
-        rm_vectors.append("slu_v_grow")
-        # TODO: include in new fn clean_method_3
-        grass.run_command(
-            "r.slope.aspect",
-            elevation=dem,
-            aspect="aspect_slu",
-            quiet=True,
-        )
-        rm_rasters.append("aspect_slu")
-        # os.system(
-        #    "/home/alvioli/dem_globo_srtm/range/script04/"
-        #    "slu_code/clean_method_3.sh slu_v_grow vect2 %s"
-        #    % cleansize
-        # )
-        clean_method_3("slu_v_grow", "vect2", cleansize)
-
-        # applying method 2 at the end
-        grass.run_command(
-            "v.to.rast",
-            input="vect2",
-            output="rast2",
-            use="cat",
-            quiet=True,
-        )
-        rm_rasters.append("rast2")
-
-        clean_input = "rast2"
-        clean_output = "slu_r_grow"
-        grass.run_command(
-            "r.neighbors",
-            input=clean_input,
-            output="slu_diversity",
-            method="diversity",
-            size=5,
-            quiet=True,
-        )
-        rm_rasters.append("slu_diversity")
-
-        exp = "$out = if($a == 1, 1, null())"
-        grass.mapcalc(
-            exp,
-            out="slu_diversity_nobordi",
-            a="slu_diversity",
-            quiet=True,
-        )
-        rm_rasters.append("slu_diversity_nobordi")
-
-        grass.run_command(
-            "r.grow",
-            input="slu_diversity_nobordi",
-            output="slu_diversity_nobordi_grow",
-            radius=1.01,
-            quiet=True,
-        )
-        rm_rasters.append("slu_diversity_nobordi_grow")
-
-        exp = "$out = if(isnull($a), null(), $b)"
-        grass.mapcalc(
-            exp,
-            out="slu_finale_nobordi",
-            a="slu_diversity_nobordi_grow",
-            b=clean_input,
-            quiet=True,
-        )
-        rm_rasters.append("slu_finale_nobordi")
-
-        grass.run_command(
-            "r.grow",
-            input="slu_finale_nobordi",
-            output=clean_output,
-            radius=1000,
-            quiet=True,
-        )
-        rm_rasters.append(clean_output)
-
-        exp = "$out = int($a)"
-        # output="slu_r_grow" exists already
-        grass.run_command(
-            "g.remove", type="raster", name=clean_input, flags="f", quiet=True
-        )
-        grass.mapcalc(exp, out=clean_input, a=clean_output, quiet=True)
-        rm_rasters.append(clean_input)
-
-        grass.run_command(
-            "g.remove",
-            type="raster",
-            name=(
-                "slu_diversity,slu_diversity_nobordi,"
-                "slu_diversity_nobordi_grow,slu_finale_nobordi,rast2"
-            ),
-            flags="f",
-            quiet=True,
-        )
-
-        grass.run_command(
-            "g.remove",
-            type="vector",
-            name="slu_v_grow,vect2",
-            flags="f",
-            quiet=True,
-        )
-
-    if options["plainsmap"]:
-        exp = "$out = if(isnull($b), if(isnull($c), null(), int($a)), null())"
-        grass.mapcalc(
-            exp,
-            out=slumapclean,
-            a="slu_r_grow",
-            b=plains,
-            c=dem,
-            quiet=True,
-        )
-    else:
-        exp = "$out = if(isnull($c), null(), int($a))"
-        grass.mapcalc(exp, out=slumapclean, a="slu_r_grow", c=dem, quiet=True)
-
-    grass.run_command("r.colors", map=slumapclean, color="random", quiet=True)
-
-    grass.message("Cleaning of small areas finished.")
-
-
 def main():
     """Main function of r.slopeunits"""
     global rm_rasters, rm_vectors
@@ -1465,22 +718,6 @@ def main():
     red = int(options["rf"])
     maxiter = int(options["maxiteration"])
 
-    cleansize = -1
-    slumapclean = None
-    if options["cleansize"]:
-        cleansize = int(options["cleansize"])
-        if options["slumapclean"]:
-            slumapclean = options["slumapclean"]
-        else:
-            grass.fatal(
-                "When cleansize is provided, slumapclean is mandatory."
-            )
-        if flags["m"] and flags["n"]:
-            grass.fatal(
-                "When cleansize is provided, only one between m and n can be "
-                "specified."
-            )
-
     slope_units(
         dem,
         plains,
@@ -1493,10 +730,6 @@ def main():
         red,
         maxiter,
     )
-
-    if cleansize > 0:
-        clean_small_areas(dem, slumap, plains, cleansize, slumapclean)
-
     grass.message("Slope units finished.")
